Transition from question modal to cards
Reuse tutorial cards for interrogative debugging questions. These can be moved
around and closed, and other fancy stuff.
Replaces the usage of modals.

These changes follow the design guidelines of React, unlike the previous version.
The used virtual machine instance is NOT pushed down to components anymore, but
handled in a dedicated container. Now reducers are used, etc. As it should be.

When a statement is highlighted, users can now click on it and the block should glow.

Many many other changes too, but more or less just improvement of the old version,
both concept and content.